64-bit prep part 1: Processed source .h and .m files using /Developer/Extras/64BitConversion/ConvertCocoa64. Worked through the source once to check for incorrect changes. We have to check the "#warning 64BIT" macro locations to complete this step of preparing for 64-bit.

+++ Audio/AudioScheduler.h
@@ -33,8 +33,8 @@ extern NSString * const		ScheduledAudioRegionObjectKey;		// ScheduledAudioRegion
 
 @interface AudioScheduler : NSObject
 {
-	unsigned				_numberSlices;
-	unsigned				_framesPerSlice;
+	NSUInteger				_numberSlices;
+	NSUInteger				_framesPerSlice;
 	
 	AudioUnit				_audioUnit;
 	
@@ -56,8 +56,8 @@ extern NSString * const		ScheduledAudioRegionObjectKey;		// ScheduledAudioRegion
 }
 
 // Buffer size information (set at object creation)
-- (unsigned) numberOfSlicesInBuffer;
-- (unsigned) numberOfFramesPerSlice;
+- (NSUInteger) numberOfSlicesInBuffer;
+- (NSUInteger) numberOfFramesPerSlice;
 
 // The ScheduledSoundPlayer AudioUnit on which to schedule audio slices
 - (AudioUnit) audioUnit;
